Migrate Sales family + fix structural-hook regression on Catalog
Sales family migrated to v5 split-class layout:

  OrderResource           267 -> 102 lines (+OrderTable 152, no form — modal-edited)
  CustomerResource        307 -> 119 lines (+CustomerForm 130, +CustomerTable 90)
  CustomerGroupResource   152 -> 65 lines  (+CustomerGroupForm 80, +CustomerGroupTable 55)

Cross-resource consumers updated in lockstep:
  - LatestOrdersTable widget → OrderTable::getColumns()
  - CustomerResource\RelationManagers\OrdersRelationManager → OrderTable::getColumns()

Also fixes a regression introduced in the Catalog PR (#2481): migrated
resources overrode public getRelations() / getPages() /
getRecordSubNavigation() directly, bypassing the BaseResource wrapper
traits and silencing the structural extension hooks (getRelations,
extendPages, extendSubNavigation). These hooks have no Form/Table
seam — they belong on the Resource — so the right pattern is to
override the protected getDefault* slots and let the trait keep its
hook indirection.

Pattern is now:
  - form() / table()  defined directly, delegate to split class
                      (hook target = {Resource}Form / {Resource}Table)
  - getDefaultPages / getDefaultRelations / getDefaultSubNavigation
                      override the protected slot; trait fires the
                      hook against the Resource

All 9 Catalog resources retrofitted to this pattern. ResourceExtensionTest
'can extend relationship managers' now passes again on CustomerResource.
