+++ packages/admin/src/Filament/Resources/CustomerGroupResource/Schemas/CustomerGroupForm.php
@@ -1,0 +1,76 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\CustomerGroupResource\Schemas;
+
+use Filament\Forms\Components\TextInput;
+use Filament\Forms\Components\Toggle;
+use Filament\Schemas\Components\Component;
+use Filament\Schemas\Components\Section;
+use Filament\Schemas\Components\Utilities\Set;
+use Filament\Schemas\Schema;
+use Illuminate\Support\Str;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+use Lunar\Admin\Support\Forms\Components\Attributes;
+
+class CustomerGroupForm
+{
+    use CallsHooks;
+
+    public static function configure(Schema $schema): Schema
+    {
+        return self::callStaticLunarHook(
+            'configureForm',
+            $schema->components([
+                Section::make()->schema(static::getMainComponents()),
+            ]),
+        );
+    }
+
+    public static function getMainComponents(): array
+    {
+        return [
+            static::getNameComponent(),
+            static::getHandleComponent(),
+            static::getDefaultComponent(),
+            static::getAttributeDataComponent(),
+        ];
+    }
+
+    public static function getNameComponent(): Component
+    {
+        return TextInput::make('name')
+            ->label(__('lunarpanel::customergroup.form.name.label'))
+            ->required()
+            ->maxLength(255)
+            ->autofocus();
+    }
+
+    public static function getHandleComponent(): Component
+    {
+        return TextInput::make('handle')
+            ->label(__('lunarpanel::customergroup.form.handle.label'))
+            ->required()
+            ->unique(ignoreRecord: true)
+            ->live(onBlur: true)
+            ->afterStateUpdated(function (string $operation, $state, Set $set) {
+                if ($operation !== 'create') {
+                    return;
+                }
+
+                $set('handle', Str::snake(Str::lower($state)));
+            })
+            ->minLength(3)
+            ->maxLength(255);
+    }
+
+    public static function getDefaultComponent(): Component
+    {
+        return Toggle::make('default')
+            ->label(__('lunarpanel::customergroup.form.default.label'));
+    }
+
+    public static function getAttributeDataComponent(): Component
+    {
+        return Attributes::make();
+    }
+}
