+++ packages/admin/src/Filament/Resources/ProductTypeResource.php
@@ -52,12 +52,7 @@ class ProductTypeResource extends BaseResource
         return ProductTypeTable::configure($table);
     }
 
-    public static function getRelations(): array
-    {
-        return [];
-    }
-
-    public static function getPages(): array
+    protected static function getDefaultPages(): array
     {
         return [
             'index' => ListProductTypes::route('/'),
