+++ packages/admin/src/Filament/Widgets/Dashboard/Orders/LatestOrdersTable.php
@@ -4,7 +4,7 @@ namespace Lunar\Admin\Filament\Widgets\Dashboard\Orders;
 
 use Filament\Tables\Table;
 use Filament\Widgets\TableWidget;
-use Lunar\Admin\Filament\Resources\OrderResource;
+use Lunar\Admin\Filament\Resources\OrderResource\Tables\OrderTable;
 use Lunar\Core\Models\Order;
 
 class LatestOrdersTable extends TableWidget
@@ -16,7 +16,7 @@ class LatestOrdersTable extends TableWidget
         return $table->query(function () {
             return Order::with(['currency'])->orderBy('placed_at', 'desc')->orderBy('created_at', 'desc')->limit(10);
         })->columns(
-            OrderResource::getTableColumns()
+            OrderTable::getColumns()
         )->paginated(false)->searchable(false)
             ->heading(__('lunarpanel::widgets.dashboard.orders.latest_orders.heading'))
             ->poll('60s');
